cdk/requirements.txt
# AWS CDK v2 dependencies for AegisClaims AI Platform
aws-cdk-lib>=2.120.0
constructs>=10.0.0

# Additional CDK constructs
aws-cdk.aws-lambda-python-alpha>=2.120.0a0

# Python utilities
python-dotenv>=1.0.0
